Cleanup metrics classes + bug fixes (#271)
Integ tests are failing, debugging them independently but they don't look related to the changes in this PR
