+++ sycamore/evaluation/__init__.py
@@ -1,4 +1,3 @@
-from sycamore.evaluation.data import EvaluationDataPoint
-from sycamore.evaluation.metrics import EvaluationMetric, document_retrieval_metrics, generated_answer_metrics
+from sycamore.evaluation.data import EvaluationDataPoint, EvaluationMetric
 
-__all__ = ["EvaluationMetric", "EvaluationDataPoint", "document_retrieval_metrics", "generated_answer_metrics"]
+__all__ = ["EvaluationDataPoint", "EvaluationMetric"]
