+++ sycamore/evaluation/metrics/__init__.py
@@ -1,0 +1,4 @@
+from sycamore.evaluation.metrics.generated_answer import rouge_metrics
+from sycamore.evaluation.metrics.retrieval import document_retrieval_metrics
+
+__all__ = ["document_retrieval_metrics", "rouge_metrics"]
