+++ sycamore/functions/tokenizer.py
@@ -1,9 +1,18 @@
+import tiktoken
 from transformers import AutoTokenizer
 
 
 class Tokenizer:
     def tokenize(self, text: str):
         pass
+
+
+class OpenAITokenizer(Tokenizer):
+    def __init__(self, model_name: str):
+        self._tk = tiktoken.encoding_for_model(model_name)
+
+    def tokenize(self, text: str):
+        return self._tk.encode(text)
 
 
 class CharacterTokenizer(Tokenizer):
